requirements.txt
Flask
Flask-Restful
flask-jwt-extended
flask-testing
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
requests
sendgrid
python-dotenv
pylint